Massive Update - see log message.
Debug mode enabled. Need to reprovision VM if it is already provisioned. Running the LabyrinthServer script with -d or --debug will trigger the debug mode. Can also manually run 'gradle runDebug'. Will setup the server to listen for a debugger on port 5005. Will wait for the debugger.

Fixed issue with Invalid Data exceptions printing nulls.

Added a status code for server errors.

Stopped the server if we cannot set it up, which without that would cause the sever UI to end up in a detached state when you couldn't do anything but Ctrl-C out of it.

Fixed event registration so it uses the socket event name not the method name