MASSIVE MASSIVE UPDATE. (Jeez these things are frequent).
So now a connection service is a pluggable thing - if we ever need a different way to connect a player (e.g Java RMI), we can make new a new class that implements the ClientConnectionService and a new Client type that implements GameClient, and we are good to go!

The system now includes bots (untested) that are able to connect and act as players. The base class is GameBot, and that handles registration with the system for you. Extend from that to make a new bot type.

Very minor updates to matches to start getting them working with bots.